Feature/fix

+++ src/main/java/com/hororok/monta/entity/Item.java
@@ -1,12 +1,10 @@
 package com.hororok.monta.entity;
 
-import com.hororok.monta.dto.request.item.PatchItemRequestDto;
 import com.hororok.monta.dto.request.item.PostItemRequestDto;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.Where;
 
@@ -24,7 +22,7 @@ public class Item extends CommonEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "item_id")
-    private long id;
+    private int id;
 
     @NotNull
     @Column(length = 20)
