+++ src/main/java/com/hororok/monta/repository/ItemRepository.java
@@ -2,10 +2,9 @@ package com.hororok.monta.repository;
 
 import com.hororok.monta.entity.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.Optional;
 
-public interface ItemRepository extends JpaRepository<Item, Long> {
+public interface ItemRepository extends JpaRepository<Item, Integer> {
 
-    Optional<Item> findOneById(Long itemId);
+    Optional<Item> findOneById(int itemId);
 }
